인터페이스 및 추상 클래스 생성 Fixes #2
