+++ client/src/core/interfaces/GameObject.ts
@@ -1,5 +1,7 @@
+import { GameContext } from "./CoreContext";
+
 export abstract class GameObject {
     abstract update(elapsed: number): void;
-    abstract draw(ctx: CanvasRenderingContext2D): void;
+    abstract draw(ctx: GameContext): void;
     abstract destroy(): void;
 }
