+++ client/src/core/interfaces/IGameState.ts
@@ -1,8 +1,9 @@
+import { GameContext } from "./CoreContext";
 import { GameObject } from "./GameObject";
 
 export interface IGameState {
     update(elapsed: number): void;
-    render(ctx: CanvasRenderingContext2D): void;
+    render(ctx: GameContext): void;
     onEnter(): boolean;
     onExit(): boolean;
