+++ client/src/core/interfaces/ISprite.ts
@@ -1,0 +1,16 @@
+import { GameObject } from "./GameObject";
+import { SpriteData } from "../sprite-data";
+
+export interface ISprite extends GameObject {
+    initialize(
+        x: number,
+        y: number,
+        width: number,
+        height: number,
+        maxFrames: number,
+        textuereId: string
+    ): boolean;
+    update(elapsed: number): void;
+
+    getSpriteData(): SpriteData;
+}
